pyproject.toml
[project]
name = "seminario-db"
description = "Gestión de recursos y reservas para seminario usando OracleDB"
authors = [{ name = "Andres Martin", email = "felipemartinrod@gmail.com" }]
version = "0.1.0"
requires-python = ">=3.10"

[tool.poetry.dependencies]
python = ">=3.10"
sqlalchemy = "^2.0"
oracledb = "^1.4"
python-dotenv = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^24.0"
ruff = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
